CustomFileLoader: refactor and decompose
