+++ src/components/@ui/CustomFields/CustomFileLoader/hooks/index.js
@@ -1,0 +1,7 @@
+import useFileValidation from "./useFileValidation";
+import useFileLoader from "./useFileLoader";
+
+export {
+  useFileValidation,
+  useFileLoader,
+};
